Refactor: Extract common GitHub data update logic
This commit extracts the `getSha` and `_updateData` methods from `DataUpdateInfo` into a new `CommonsServices` class.

This change aims to centralize common GitHub interaction logic, making `DataUpdateInfo` more focused on its specific data update responsibilities and improving code organization.

Specific changes:
- Created `lib/core/services/commons/commons_services.dart` with the `CommonsServices` class.
- Moved `getSha` and `_updateData` (renamed to `updateData`) methods to `CommonsServices`.
- Updated `DataUpdateInfo` to use `CommonsServices` for GitHub operations.
- Removed unused imports from `DataUpdateInfo`.
